Stop loading editor and media scripts on the front end
utils.js was shared between the editor and front-end bundles but held two
editor-only helpers, so the dependency extraction plugin recorded
wp-block-editor in view.asset.php and WordPress pulled in wp-media-utils
and the media scripts wherever a dynamic shape block rendered.

Move useSpacingPresets and axisConfig to editor-utils.js, leaving utils.js
free of editor imports. view.asset.php now declares only wp-dom-ready.
